Add kernel config hound
Add a checker for CHROMIUM patches that looks to see if anything
in chromeos/config/* is in the CL. If it is, then the new
KernelConfigChecker will report the difference of all possible
configs for each board before and after the CL.

The goal of this is to mitigate any potential risk of the new
savedefconfig kernel configs (proposed at go/mini-splitconfigs)
by spewing onto the CL the total difference in config changes,
which might not be visible in the diff due to the now minimized
savedefconfigs.

Once such instance of this that we might like to see is a recent
change involving the MediaTek pinctrl drivers. Flipping on mediatek
in general defaulted the pinctrl drivers for all SoCs to Y. This
SeanBot change would make it more obvious that by enabling mediatek,
we were now including oodles of pinctrl drivers.

I'm not totally sure how this will be received to people other than
Doug, so I've added a command line argument --kconfig-hound, to
enable this checker. The other thing this checker does which is new
is that it actually utilizes the kernel working directory. If this
is disruptive, I think I could silo this off into its own temporary
working directory.

Signed-off-by: Evan Green <evangreen86@gmail.com>