Make updating remotes more robust
Update the remote's url every time we try to fetch it instead of just
assuming it's correct (which is a _pretty_ reasonable assumption since
remote names are generate from the url, but..).

This allows us a bit more fine-grained error handling since now git add
shouldn't fail and we can bounce an error. Likewise, we had to remove
the remote if fetch failed since we didn't know if the url was correct.
Now we know the remote will be updated the next time around and can keep
it around.

Keeping the remote around on fetch failures makes us less susceptible to
false negatives when the network is flakey.

Signed-off-by: Sean Paul <seanpaul@chromium.org>