FETCH_HEAD Considered Harmful
I think this will solve Issue #19.

The problem is that if an unrelated git fetch is done in the review repo
(like if someone is running git fetch in a cron job because it speeds
things up) in between

git fetch GERRIT_REF
<---------------------- git fetch cros
git show FETCH_HEAD

FETCH_HEAD will no longer point to the fetched GERRIT_REF, instead pointing
to whatever was fetched last. To fix this, fetch GERRIT_REF into a named
temporary ref and use that temporary ref to show the patch.

In the case of the missing hash reviews, all of these races ended up
doing a review on the diff from http://crosreview.com/2630 (probably
because the branch name is alphabetically first), which has a missing
hash.

Signed-off-by: Sean Paul <seanpaul@chromium.org>