Change requirements on patchwork urls
Not all patchwork urls end with a numeric id, so treat whatever comes
after '/patch/' as the id.

This fixes patches such as: http://crosreview.com/1454878

The downside to using msgid is that the api doesn't recognize msgid so
we can get comments or series without parsing the id first. Maybe some
other time..

Signed-off-by: Sean Paul <seanpaul@chromium.org>