troll: Process retry requests through topics
Gerrit now embeds top-level messages as comments which are not returned
from patch queries. As such, we can no longer process retry requests
as messages since this would require us to fetch comments for each
change in a query result.

Instead, use the change topic to initiate a retry. This is a bit simpler
since we can clear the topic and avoid having to compare last review
date with the retry request.

Signed-off-by: Sean Paul <seanpaul@chromium.org>