troll: Increase diff indentation to 4 spaces
Gerrit previously supported a subset of markdown features and accepted
lines prefixed with two spaces as monospaced formatting.

Recently Gerrit has added more robust support for markdown more in line
with the spec, which defines 4 leading spaces for monospaced formatting:
https://commonmark.org/help/

Increase the leading spaces for the diff output to ensure it is
formatted correctly on Gerrit.

Signed-off-by: Drew Davenport <ddavenport@chromium.org>